Some added

+++ src/main/java/com/softserve/task4/models/Man.java
@@ -1,7 +1,7 @@
 package com.softserve.task4.models;
 
 public class Man extends Human {
-    public Man(String firstName, String lastName, float height, float weight) {
+    public Man(String firstName, String lastName, double height, double weight) {
         super(true, firstName, lastName, height, weight);
     }
